Arguments Added and Updated
A new argument called -z or --noinstall has been added to allow the user to choose that the program not try to install the packages it needs, which is helpful for super computers where the user doesn't have write access to where python is stored. This WILL require the user manually install all the other packages in requirements themselves though. 4 other arguments were updated as well so that they only need to be called to be toggled on/off, which made a package (str2bool) irrelevant and no longer required. The requirements.txt and README has been updated to reflect these changes as well.